Update main.py
Ajoute un système de classement qui s'active pour la StratAcademy uniquement le temps de tester
Réduction du nombre de requête SQL au global
Augmentation de la rapidité du bot au global

+++ main.py
@@ -1,14 +1,19 @@
+import datetime
+from zoneinfo import ZoneInfo
+
 import discord
 import requests
 from discord.ext import tasks, commands
+
 from Database import Database
 
 compteur = 0
-
+db = Database()
+admin_atlas_id = 114372984604590080
+stratacademy_id = 799713691447853057
 riot_api_key = "X"
 TOKEN = "X"
 
-db = Database()
 intents = discord.Intents.all()
 client = commands.Bot(command_prefix="/", intents=intents)
 
@@ -31,12 +36,11 @@ ranks = {
 }
 
 
-def displayInfo(player, guildid):
-    refreshedPlayer = db.recoverSpecificJoueur(player[0], guildid)
-    temp = "Le joueur " + refreshedPlayer[1] + " est classé " + str(refreshedPlayer[2]) + " " + \
-           str(refreshedPlayer[3]) + " avec " + str(refreshedPlayer[4]) + " LPs."
-    if refreshedPlayer[5] == 1:
-        x = refreshedPlayer[6].replace('W', ":white_check_mark: ")\
+def displayInfo(player):
+    temp = "Le joueur " + player.get("summonername") + " est classé " + str(player.get("tier")) + " " + \
+           str(player.get("rank")) + " avec " + str(player.get("lps")) + " LPs."
+    if player.get("enBo") == 1:
+        x = player.get("progress").replace('W', ":white_check_mark: ")\
             .replace('L', ":no_entry_sign: ").replace('N', ":clock3: ")
         temp += "\nLe joueur est actuellement en BO : " + x
     return temp
@@ -54,6 +58,7 @@ def createPlayer(acc, rank, guild, member_id):
                 enBo = False
             rc = db.addJoueur(acc.get('id'), acc.get('name'), typegames.get('tier'), typegames.get('rank'),
                               typegames.get('leaguePoints'), enBo, progress, guild.id, member_id)
+            db.AddClassement(acc.get('id'))
             return rc
 
 
@@ -89,7 +94,7 @@ def check_rang(player, guild):
     for typequeue in ranking:
         if typequeue.get('queueType') == 'RANKED_SOLO_5x5':
             ret = ""
-            atmelo = {
+            eloactuel = {
                 "summonername": player[1],
                 "tier": player[2],
                 "rank": player[3],
@@ -97,66 +102,84 @@ def check_rang(player, guild):
                 "enBo": player[5],
                 "progress": player[6]
             }
+            newelo = {
+                "summonername": typequeue.get('summonerName'),
+                "tier": typequeue.get('tier'),
+                "rank": typequeue.get('rank'),
+                "lps": typequeue.get('leaguePoints'),
+                "enBo": 0,
+                "progress": None
+            }
             if guild[3] != 0:
                 ret += "<@&" + str(guild[3]) + "> "
             if typequeue.get('miniSeries') is not None:
-                atmelo["enBo"] = True
+                eloactuel["enBo"] = True
+                newelo["enBo"] = 1
                 miniS = typequeue.get('miniSeries')
                 progress = miniS.get('progress')
-                if progress != atmelo["progress"]:
+                newelo["progress"] = progress
+                if progress != eloactuel["progress"]:
                     db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'),
                                     typequeue.get('rank'),
-                                    typequeue.get('leaguePoints'), atmelo.get("enBo"), progress)
-                    return ret
+                                    typequeue.get('leaguePoints'), eloactuel.get("enBo"), progress)
+                    tab_progress = list(filter('N'.__ne__, progress))
+                    if len(tab_progress) != 0 and tab_progress[len(tab_progress) - 1] == 'W':
+                        db.UpdateWinClassement(player[0])
+                    return [ret, newelo]
             else:
-                atmelo["enBo"] = False
-            if typequeue.get('tier') == atmelo.get("tier") and typequeue.get('rank') == atmelo.get("rank") and \
-                    typequeue.get('leaguePoints') == atmelo.get("lps"):
-                return "RAS"
-            if typequeue.get('tier') != atmelo.get("tier") \
-                    and tiers.get(typequeue.get('tier')) < tiers.get(atmelo.get("tier")):
-                ret += str(typequeue.get('summonerName')) + " a derank de " + atmelo.get('tier') + " à " \
+                eloactuel["enBo"] = False
+            if typequeue.get('tier') == eloactuel.get("tier") and typequeue.get('rank') == eloactuel.get("rank") and \
+                    typequeue.get('leaguePoints') == eloactuel.get("lps"):
+                return ["RAS", newelo]
+            if typequeue.get('tier') != eloactuel.get("tier") \
+                    and tiers.get(typequeue.get('tier')) < tiers.get(eloactuel.get("tier")):
+                ret += str(typequeue.get('summonerName')) + " a derank de " + eloactuel.get('tier') + " à " \
                        + typequeue.get('tier')
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
-            if typequeue.get('tier') != atmelo.get("tier") \
-                    and tiers.get(typequeue.get('tier')) > tiers.get(atmelo.get("tier")):
-                ret += str(typequeue.get('summonerName')) + " a rank up de " + atmelo.get('tier') + " à " \
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                return [ret, newelo]
+            if typequeue.get('tier') != eloactuel.get("tier") \
+                    and tiers.get(typequeue.get('tier')) > tiers.get(eloactuel.get("tier")):
+                ret += str(typequeue.get('summonerName')) + " a rank up de " + eloactuel.get('tier') + " à " \
                        + typequeue.get('tier')
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                db.UpdateWinClassement(player[0])
+                return [ret, newelo]
 
-            if typequeue.get('rank') != atmelo.get("rank") \
-                    and ranks.get(typequeue.get('rank')) < ranks.get(atmelo.get("rank")):
-                ret += str(typequeue.get('summonerName')) + " est descendu de " + atmelo.get("tier") + ' ' + \
-                       atmelo.get("rank") \
+            if typequeue.get('rank') != eloactuel.get("rank") \
+                    and ranks.get(typequeue.get('rank')) < ranks.get(eloactuel.get("rank")):
+                ret += str(typequeue.get('summonerName')) + " est descendu de " + eloactuel.get("tier") + ' ' + \
+                       eloactuel.get("rank") \
                        + " à " + typequeue.get('tier') + ' ' + typequeue.get('rank')
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
-            if typequeue.get('rank') != atmelo.get("rank") \
-                    and ranks.get(typequeue.get('rank')) > ranks.get(atmelo.get("rank")):
-                ret += str(typequeue.get('summonerName')) + " est monté de " + atmelo.get("tier") + ' ' + \
-                       atmelo.get("rank") \
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                return [ret, newelo]
+            if typequeue.get('rank') != eloactuel.get("rank") \
+                    and ranks.get(typequeue.get('rank')) > ranks.get(eloactuel.get("rank")):
+                ret += str(typequeue.get('summonerName')) + " est monté de " + eloactuel.get("tier") + ' ' + \
+                       eloactuel.get("rank") \
                        + " à " + typequeue.get('tier') + ' ' + typequeue.get('rank')
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                db.UpdateWinClassement(player[0])
+                return [ret, newelo]
 
-            if typequeue.get('leaguePoints') != atmelo.get("lps") and typequeue.get('leaguePoints') < atmelo.get("lps"):
+            if typequeue.get('leaguePoints') != eloactuel.get("lps") and typequeue.get('leaguePoints') < \
+                    eloactuel.get("lps"):
                 ret += str(typequeue.get('summonerName')) + " a perdu -" + \
-                       str(atmelo.get("lps") - typequeue.get('leaguePoints')) + " LPs"
+                       str(eloactuel.get("lps") - typequeue.get('leaguePoints')) + " LPs"
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
-            if typequeue.get('leaguePoints') != atmelo.get("lps") and typequeue.get('leaguePoints') > atmelo.get("lps"):
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                return [ret, newelo]
+            if typequeue.get('leaguePoints') != eloactuel.get("lps") and typequeue.get('leaguePoints') > \
+                    eloactuel.get("lps"):
                 ret += str(typequeue.get('summonerName')) + " a gagné +" + \
-                       str(typequeue.get('leaguePoints') - atmelo.get("lps")) + " LPs"
+                       str(typequeue.get('leaguePoints') - eloactuel.get("lps")) + " LPs"
                 db.updateJoueur(player[0], typequeue.get('summonerName'), typequeue.get('tier'), typequeue.get('rank'),
-                                typequeue.get('leaguePoints'), atmelo.get("enBo"), atmelo.get("progress"))
-                return ret
+                                typequeue.get('leaguePoints'), eloactuel.get("enBo"), eloactuel.get("progress"))
+                db.UpdateWinClassement(player[0])
+                return [ret, newelo]
 
 
 @client.event
@@ -176,24 +199,33 @@ async def on_ready():
     print("Chargement de l'activité...")
     await client.change_presence(activity=discord.Game(name="Je vous vois tous"))
     on_update.start()
+    classement.start()
 
 
 @client.event
 async def on_guild_join(guild):
+    print("GuildAdd : le bot a été ajouté sur un serveur")
     db.addServeur(guild.id, guild.name, 0, 0)
     return
 
 
 @client.event
 async def on_guild_remove(guild):
+    print("GuildRemove : un serveur a supprimé le bot")
     db.removeServeur(guild.id)
     db.removeAllJoueurs(guild.id)
     return
 
 
 @client.event
-async def on_raw_member_remove(payload):
-    db.RemoveJoueur(payload.guild_id, payload.user.id)
+async def on_member_remove(member):
+    print("MemberRemove : un joueur a quitté un serveur qui a entraîné son retrait de la BDD")
+    rowCount, res = db.GetJoueurFromMemberId(member.id)
+    if rowCount == 1:
+        db.RemoveJoueur(member.guild.id, member.id)
+        db.DeleteClassement(res[0][0])
+    elif rowCount > 1:
+        db.RemoveJoueur(member.guild.id, member.id)
     return
 
 
@@ -207,6 +239,7 @@ async def on_message(message):
 @client.tree.command(name="initialize", description="Initialise un nouveau serveur")
 async def initialize(ints, channelmessage: discord.TextChannel,
                      roleaping: discord.Role = None):
+    print("Initialisation : une initialisation a été demandée")
     try:
         msg = "Message test. Si vous voyez ce message, cela signifie que le bot a l'autorisation d'écrire dans le " \
               "channel. Vous pouvez le supprimer dès la fin de l'initialisation."
@@ -227,6 +260,7 @@ async def initialize(ints, channelmessage: discord.TextChannel,
 
 @client.tree.command(name="addjoueur", description="S'ajouter dans la liste des joueurs")
 async def addJoueur(ints, nomjoueur: str):
+    print("Addjoueur : une demande d'ajout a été envoyée")
     ret = addPlayer(nomjoueur, ints.guild, ints.user.id)
     if ret is None:
         msg = "Erreur lors de l'ajout du joueur. Veuillez vérifier qu'il existe bien, qu'il est niveau 30 et qu'il" \
@@ -242,8 +276,16 @@ async def addJoueur(ints, nomjoueur: str):
 
 @client.tree.command(name="leavelolwatcher", description="Se retirer de la liste des joueurs")
 async def removeJoueur(ints):
-    res = db.RemoveJoueur(ints.guild_id, ints.user.id)
-    if res == 1:
+    print("LeaveLoLWatcher : une demande de retrait a été envoyée")
+    rowCount, res = db.GetJoueurFromMemberId(ints.user.id)
+    if rowCount == 1:
+        db.DeleteClassement(res[0][0])
+    exist = False
+    for joueur in res:
+        if joueur[7] == ints.guild_id:
+            exist = True
+    if exist:
+        db.RemoveJoueur(ints.guild_id, ints.user.id)
         await ints.response.send_message("Vous avez été retiré de la liste")
     else:
         await ints.response.send_message("Vous n'êtes pas présent dans la liste")
@@ -252,15 +294,19 @@ async def removeJoueur(ints):
 @client.tree.command(name="listejoueurs", description="Liste des joueurs")
 async def listeJoueurs(ints):
     await ints.response.defer()
+    print("Listejoueurs : une liste a été demandée à la BDD")
     res = db.GetJoueursOfGuild(ints.guild_id)
     g = ints.guild
     retour = "Liste de(s) joueur(s) : \n"
     if not res:
         retour = "La liste des joueurs est vide"
     else:
-        for i in db.GetJoueursOfGuild(ints.guild_id):
-            m = await g.fetch_member(i[8])
-            retour += " - " + i[1] + " (" + m.name + ") \n"
+        for i in res:
+            try:
+                m = await g.fetch_member(i[8])
+                retour += " - " + i[1] + " (" + m.name + ") \n"
+            except discord.NotFound:
+                retour += " - " + i[1] + " (ERREUR : NOT FOUND) \n"
         temp = retour.rsplit('\n', 1)
         retour = ''.join(temp)
     await ints.followup.send(retour)
@@ -269,6 +315,7 @@ async def listeJoueurs(ints):
 @client.tree.command(name="infojoueur", description="Donne les infos d'un joueur")
 async def info_joueur(ints, summonername: str):
     await ints.response.defer()
+    print("Infojoueur : une info de joueur a été demandée à la BDD")
     p = db.GetPlayerInfo(ints.guild_id, summonername)
     if p:
         temp = "Le joueur " + p[1] + " est classé " + str(p[2]) + " " + \
@@ -281,10 +328,10 @@ async def info_joueur(ints, summonername: str):
         temp = "Erreur lors de la récupération du joueur. Veuillez vérifier qu'il existe bien"
         await ints.followup.send(temp)
 
-
 @client.tree.command(name="infojoueurdiscord", description="Donne les infos d'un joueur")
 async def info_joueur_discord(ints, membre: discord.Member):
     await ints.response.defer()
+    print("Infojoueurdiscord : une info de membre Discord a été demandée à la BDD")
     p = db.GetPlayerInfoDiscord(ints.guild_id, membre.id)
     if p:
         temp = "Le joueur " + p[1] + " est classé " + str(p[2]) + " " + \
@@ -310,6 +357,7 @@ async def alertGuilds(ints, message: str):
             await channel.send("Message de l'admin : \n>>> " + message)
         except Exception as e:
             print(e)
+    print("Alert : une alerte a été envoyée aux serveurs")
     await ints.followup.send("L'alerte a bien été envoyée")
 
 
@@ -317,7 +365,7 @@ async def alertGuilds(ints, message: str):
 async def alert_admin(ints, message: str):
     await ints.response.defer()
     atlas = await client.fetch_user(admin_atlas_id)
-    ret = "<@" + str(ints.user.id) + "> vous a envoyé un meesage : \n\n" + message
+    ret = "<@" + str(ints.user.id) + "> vous a envoyé un message : \n\n" + message
     await atlas.send(ret)
     await ints.followup.send("Votre message a bien été envoyé. Vous serez recontacté sous peu."
                              " Merci de ne pas spam la commande")
@@ -327,22 +375,36 @@ async def alert_admin(ints, message: str):
 async def on_update():
     global compteur
     compteur += 1
-    print("Vérification n°" + str(compteur))
-    for g in db.recoverAllGuilds():
-        channel = client.get_channel(g[2])
-        for i in db.GetJoueursOfGuild(g[0]):
-            retour = check_rang(i, g)
-            if retour is None:
-                print("Erreur RIOT API.")
-            elif retour == "RAS":
-                print(i[1] + " n'a pas joué de partie")
-            else:
-                retour += "\n" + displayInfo(i, g[0])
-                try:
-                    await channel.send(str(retour))
-                except discord.errors.Forbidden:
-                    print("Error guild '" + g[1] + "' : Le bot n'a pas le droit d'écrire dans le channel initialisé.")
+    print("\nVérification n°" + str(compteur))
+    # ICI LA => Vérification est pas encore faite
+    for i in db.UpdatePlayerRecover():
+        channel = client.get_channel(i[11])
+        guild_infos = [i[9], i[10], i[11], i[12]]
+        retour = check_rang(i, guild_infos)
+        if retour is None:
+            print("Erreur RIOT API.")
+        elif retour[0] != "RAS":
+            retour[0] += "\n" + displayInfo(retour[1])
+            try:
+                await channel.send(str(retour[0]))
+            except discord.errors.Forbidden:
+                print("Error guild '" + guild_infos[1] + "' : Le bot n'a pas le droit d'écrire"
+                                                         " dans le channel initialisé.")
+
+
+@tasks.loop(time=datetime.time(21, 0, 0, 0, ZoneInfo("Europe/Paris")))
+async def classement():
+    td = datetime.datetime.now(ZoneInfo("Europe/Paris"))
+    if td.weekday() != 6:
+        return
+    c = db.GetClassement(stratacademy_id)
+    msg_return = "La semaine est finie ! Voici les meilleurs joueurs de la semaine : \n"
+    msg_return += ":first_place: " + c[0][0] + " (<@" + str(c[0][1]) + ">) avec " + str(c[0][2]) + " victoires !\n"
+    msg_return += ":second_place: " + c[1][0] + " (<@" + str(c[1][1]) + ">) avec " + str(c[1][2]) + " victoires !\n"
+    msg_return += ":third_place: " + c[2][0] + " (<@" + str(c[2][1]) + ">) avec " + str(c[2][2]) + " victoires !\n"
+    msg_return += "\nBravo à tous les participants. La classement est maintenant reset. A la semaine prochaine !"
+    await client.get_channel(c[0][3]).send(msg_return)
+    db.ResetClassement()
 
 
 client.run(TOKEN)
-
